defs: add big pile of in6 definitions
We will need in6_ifreq to look at interface address flags,
and that in turn pulls in a ton of additional inet6 data
structures, so add them all now.

NB: I made this match the C headers first, and only then
read that cython uses the actual C structure for underlying
code, and just uses this to import names.  So this did
not need this much work, but now that it's done, might as
well keep it.